requirement.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
alembic==1.5.8
antlr4-python3-runtime==4.8
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==20.3.0
blessings==1.7
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cliff==3.7.0
cmaes==0.8.2
cmd2==1.5.0
colorama==0.4.4
colorlog==4.8.0
command-not-found==0.3
commonmark==0.9.1
configparser==5.0.2
cryptography==2.1.4
dataclasses==0.8
distro-info==0.18ubuntu0.18.04.1
docker-pycreds==0.4.0
easydict==1.9
filelock==3.0.12
fsspec==0.8.7
future==0.18.2
gitdb==4.0.5
gitpython==3.1.14
google-auth-oauthlib==0.4.3
google-auth==1.28.0
gpustat==0.6.0
greenlet==1.0.0
grpcio==1.36.1
hydra-core==1.1.0.dev4
hydra-optuna-sweeper==1.1.0.dev1
idna-ssl==1.1.0
idna==2.10
importlib-metadata==3.10.0
importlib-resources==5.1.2
joblib==1.0.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
mako==1.1.4
markdown==3.3.4
markupsafe==1.1.1
multidict==5.1.0
netifaces==0.10.4
nltk==3.5
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
omegaconf==2.1.0.dev24
opt-einsum==3.3.0
optuna==2.4.0
packaging==20.9
pathtools==0.1.2
pbr==5.5.1
pillow==8.0.0
pip==9.0.1
prettytable==2.1.0
promise==2.3
protobuf==3.15.6
psutil==5.7.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycrypto==2.6.1
pygments==2.8.1
pygobject==3.26.1
pyparsing==2.4.7
pyperclip==1.8.2
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-distutils-extra==2.39
python-editor==1.0.4
pytorch-lightning==1.2.4
pyxdg==0.25
pyyaml==5.4.1
regex==2020.11.13
requests-oauthlib==1.3.0
requests==2.25.1
rich==9.13.0
rsa==4.7.2
sacremoses==0.0.43
scipy==1.5.4
secretstorage==2.3.1
sentry-sdk==1.0.0
setuptools==54.1.2
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
sqlalchemy==1.4.6
ssh-import-id==5.7
stevedore==3.3.0
subprocess32==3.5.4
tensorboard-plugin-wit==1.8.0
tensorboard==2.4.1
tokenizers==0.10.1
torch-struct==0.5
torch==1.7.0
torchaudio==0.7.0
torchmetrics==0.2.0
torchvision==0.8.0
tqdm==4.60.0
transformers==4.3.3
typing-extensions==3.7.4.3
ufw==0.36
urllib3==1.26.4
wandb==0.10.21
wcwidth==0.2.5
werkzeug==1.0.1
wheel==0.36.2
yarl==1.6.3
zipp==3.4.1
